Extract external session tests into DataSourceV2ExternalSessionSuite
Move cross-session tests (extSession, SharedInMemoryTableCatalog) into
a dedicated suite. These tests use different infrastructure (separate
SparkSession, shared static table map) from the rest of the suite.
